Processing step

Class containing the function and arguments to be applied in a lazy-execution framework.

Objects of this class are created using the ProcessingStep() function. The processing step is executed with the executeProcessingStep() function.

ProcessingStep(FUN = character(), ARGS = list())

executeProcessingStep(object, ...)

Arguments

FUN
function or character representing a function name.

ARGS
list of arguments to be passed along to FUN.

object
ProcessingStep object.

...
optional additional arguments to be passed along.

Details

This object contains all relevant information of a data analysis processing step, i.e. the function and all of its arguments to be applied to the data. This object is mainly used to record possible processing steps of a Spectra() object.

Examples

## Create a simple processing step object
ps <- ProcessingStep(sum)

executeProcessingStep(ps, 1:10)

#> [1] 55